Changed the pathing for the logos on the navbar and links on the Main.Master page.
Added the login and create user functionalities and respective pages.

Added a LogoutButton_Click function the Main.Master.cs page to logout the user and clear the session. And changed the navbar so that the buttons for log in, and register dissapear if the user is logged in and instead only shows a logout button