Replaced the Main.aspx with Main_page.aspx.
Added the Datalist on the Main_page.aspx with the information from the Locais that comes from the database, and added links to the Local.aspx page and sends the current Local id.

Added a link to the Search.aspx page on the Navbar in the Main.Master.

Added the search functionaly on the Search.aspx page to search by Distrito, Concelho or Nome. Also added a Datalist that dynamically updates according to the user search.