Cahnged the logout button on the Main.Master so that it doesn´t cause a validation, so that it can be clicked inside of the Add_local page.
Added checks in the add_local.aspx file so that it checks if the text boxes or the file upload are empty.

Changed the condition of the if trigger when searching only if text so that it now doesn´t crash due to not having an option selected from the Concelho datalist.

Updated the design on the create_account.aspx so that it matches the front-end look. Also changed the c# code to match the new names of the text boxes.

Deleted the old validations that were on the validation.js file.