Added the latitude and altitude feature. Also added a map for the local page to show where the location is and on the create_local page to let the user add the location on the map.
Solved the error of being logged in, reloading the project and trying to go tot the personal area by adding a check if the user is logged on, on the master.master .

changed the buttons on the create_local and eidt page so that they now cancel everything and for a option to delete the local.

Also added a button to the personal page where they can jsut click to delete the place isntead of having to do it from the edit page.