Added the profile page that allows the user change their username and email.
Added the personal area, profile, adicionar local to the navbar.

Changed the functionality to get the latitude and altitude so that it gets the data automatically trhought an api so that if the user doesn´t pick anything on the map it automatically fills the table with the data depending on the concelho that the user chose.

Added the funcionality to let the users that haven´t created any comments and aren´t the user that craeted the local, to add comments to the page.

On the create_account, Add_local, Edit_local added a set focus on the fields that were left empty.