Changed the Profile page code so that it resets the Session["id_utilizador"] if the user changed the username.
Changed the logic of the Master page so that it only changes the Session["id_utilizador"] if it is null.

Added back the functionality to add the weather for each individual local page using the ipma API.

Changed the frontend design of the main page so that the datalist and the top 10 locais are on the same row.

Changed the deisgn of all the pages so that it matches the frontend